Fiks feilende test
